requirements.txt
Authlib==1.2.1
bidict==0.22.1
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
eventlet==0.33.3
Flask==2.2.2
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
google-auth==2.23.3
google-auth-oauthlib==1.1.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
passlib==1.7.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
urllib3==2.0.7
Werkzeug==2.2.2
wsproto==1.2.0
Flask-SQLAlchemy